requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django==2.1.1
djangorestframework==3.8.2
pytz==2018.5              # via django
